requirements.txt
aiorun
aiomisc
elasticsearch-async==6.2.0
aiohttp ==3.7.4.post0
aiohttp-jinja2==1.4.2
aioresponses==0.7.2
aiohttp-retry==2.3.5
aiopg==1.2.1
jinja2==2.11.3
peewee-async==0.7.1
peewee==3.14.4
peewee_generic_filters==0.2.3
psycopg2-binary
pydantic==1.8.1
pytest-aiohttp==0.3.0

python-json-logger==2.0.1
redis==3.5.3
requests==2.25.1
uuid
uvloop == 0.15.2
asyncio-periodic

